Cleanup and modernize the repo for open source.
This updates the import paths and go.mod to reflect the repository
location change so that builds are possible and work. This also adds
some very basic github actions checks to ensure that commits actually
pass the most basic of tests. This change also removes the obsolete
circleci config that isn't used in this repo. This change also adds a
basic .gitignore that ignores the vendor directory as well as vim
swap files.

And finally internal/logger.TestLogger_Debug_NoAllocates was removed
since it doesn't work anymore and is more or less obsolete as the
internal/logging library is going to be removed in favor of the log/slog
library in the future.

Several documentation sources are updated as well: README.txt, TODO.txt.
TODO specifically was extremely obsolete so it was just removed.