Switch from internal/logging to log/slog
When Blobby was written there wasn't a really good standard for
structured logging libraries so the internal/logging library was
written and used through the code. Now, a few years later golang
has added log/slog which is a structured logging library that is
good enough, and requires a lot less explaining and maintenance.

This chane switches all of the uses of logging over, which also
necessitated adding context.Context objects to a lot of the method
calls, as wlel as plumbing to pass those values through where that
makes sense. This also uses contexts for cancelation to match the
modern golang style.