Remove the Dockerfile configs for now.
Longer term I would like to move this into a packaging directory
so that its out of the way from the main code, but for now this is
also extremely out of date and needs updated. Rather than spend
time on that I am just going to nuke and reimport later when
I have time to set that up.